+++ packages/core/src/core.js
@@ -1,0 +1,5 @@
+export default function core() {
+  // TODO
+
+  console.log('core 1231');
+}