+++ packages/example/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import * as Editor from 'editor-gun';
+import Editor from 'editor';
 // import test from '@editor/test';
+import core from 'core';
 
 const App = () => {
   const value = '来自实例的info';
@@ -8,8 +9,8 @@ const App = () => {
   return (
     <div>
       <h2>实例页面： {value}</h2>
-      <Editor />
-      {/* <p>{test()}</p> */}
+      <Editor name="big gun" address="cd" />
+      {core()}
     </div>
   );
 };