+++ packages/test/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.7](https://github.com/codeGun123/lerna-gun/compare/v1.6.5...v1.6.7) (2020-11-10)
+
+**Note:** Version bump only for package @editor/test
+
+
+
+
+
 ## [1.6.6](https://github.com/codeGun123/lerna-gun/compare/v1.6.5...v1.6.6) (2020-11-10)
 
 **Note:** Version bump only for package @editor/test